backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "A web app for automatic speech recognition using OpenAI's Whisper model running locally."
authors = ["FluxCapacitor2 <31071265+FluxCapacitor2@users.noreply.github.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[tool.poetry.dependencies]
python = "^3.12"
openai-whisper = "^20240930"
requests = "^2.32.3"
fastapi = "^0.115.0"
numba = "^0.60.0"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
python-multipart = "^0.0.11"
torch = "2.4.1+cpu"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
